@@ -3,7 +3,7 @@
   <Import Project="../core.props" />
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace></RootNamespace>
     <NoWarn>;CS0649;CS0169;CS1591;</NoWarn>
     <NukeRootDirectory>..</NukeRootDirectory>
@@ -12,8 +12,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Nuke.Common" Version="5.2.1" />
+    <PackageReference Include="Nuke.Common" Version="5.3.0" />
     <PackageReference Include="ReportGenerator" Version="4.8.*" />
-    <PackageDownload Include="GitVersion.Tool" Version="[5.6.10]" />
+    <PackageDownload Include="GitVersion.Tool" Version="[5.8.1]" />
+    <PackageDownload Include="Codecov.Tool" Version="[1.12.3]" />
   </ItemGroup>
 </Project>
