@@ -1,8 +1,9 @@
 ﻿namespace DataFilters
 {
     /// <summary>
-    /// Logic that can be apply when combining <see cref="Filter"/>s.
+    /// Logic that can be apply when combining several <see cref="Filter"/>s together.
     /// </summary>
+    /// <see cref="MultiFilter.Logic"/>
     public enum FilterLogic
     {
         /// <summary>
