@@ -104,7 +104,7 @@
                     Min?.Included,
                     Min?.Expression?.EscapedParseableString,
                     Type = Min?.Expression?.GetType().Name,
-                    DebugView = Max?.Expression?.ToString()
+                    DebugView = Min?.Expression?.ToString()
                 },
                 Max = new
                 {
@@ -117,8 +117,10 @@
             }
 #if NETSTANDARD1_3
         .Jsonify(new Newtonsoft.Json.JsonSerializerSettings() { Formatting = Newtonsoft.Json.Formatting.Indented })
-#elif NETSTANDARD2_0_OR_GREATER || NET5_0_OR_GREATER
+#elif NETSTANDARD2_0_OR_GREATER || NET5_0
         .Jsonify(new() { WriteIndented = true, IgnoreNullValues = true })
+#elif NET6_0_OR_GREATER
+        .Jsonify(new() { WriteIndented = true, DefaultIgnoreCondition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingNull})
 #endif
                 )
         ;
@@ -188,8 +190,8 @@
         /// <summary>
         /// Deconstruction method
         /// </summary>
-        /// <param name="min"></param>
-        /// <param name="max"></param>
+        /// <param name="min">lower bound of the interval</param>
+        /// <param name="max">Upper bound of the interval</param>
         public void Deconstruct(out BoundaryExpression min, out BoundaryExpression max)
         {
             min = Min;
