@@ -88,7 +88,13 @@
 #endif
 
         ///<inheritdoc />
-        public override string ToString() => $"[{Start}-{End}]";
+        public override string EscapedParseableString => $"[{Start}-{End}]";
+
+        ///<inheritdoc />
+        public override string OriginalString => $"[{Start}-{End}]";
+
+        ///<inheritdoc />
+        public override double Complexity => 1 + Math.Pow(2, End - Start);
 
         ///<inheritdoc />
         public int CompareTo(RangeBracketValue other) => (Start, End).CompareTo((other.Start, other.End));
