@@ -62,8 +62,11 @@ namespace DataFilters.Grammar.Syntax
         public virtual bool Equals(StringValueExpression other) => Equals(Value, other?.Value);
 
         ///<inheritdoc/>
-        public override bool Equals(object obj) => Equals(obj as StringValueExpression);
-
+        public override bool Equals(object obj) => obj switch {
+            StringValueExpression stringValue => Equals(stringValue),
+            ISimplifiable simplifiable => Equals(simplifiable.Simplify() as StringValueExpression),
+            _ => false
+        };
         ///<inheritdoc/>
         public override int GetHashCode() => Value.GetHashCode();
 
