@@ -3,11 +3,16 @@
     using DataFilters.Grammar.Syntax;
     using DataFilters.UnitTests.Helpers;
 
+    using FluentAssertions;
+
     using FsCheck;
     using FsCheck.Fluent;
     using FsCheck.Xunit;
 
+    using System;
     using System.Linq;
+
+    using Xunit;
 
     public class RangeBracketValueTests
     {
@@ -22,7 +27,7 @@
         }
 
         [Property]
-        public Property Given_ConstantBracketExpression_contains_consecutive_characters_Equals_should_returns_true()
+        public Property Given_ConstantBracketExpression_contains_consecutive_characters_When_comparing_to_RangeBracketValue_that_has_head_and_tail_Equals_should_returns_true()
         {
             // Arrange
             IArbMap arbMap = ArbMap.Default;
@@ -59,5 +64,34 @@
         [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
         public Property Given_left_and_right_RangeBracketValues_left_gte_right_should_be_same_as_left_gt_right_or_left_eq_right(RangeBracketValue left, RangeBracketValue right)
             => (left <= right).When(left < right || left == right).Label($"Left, Right : {(left, right)}");
+
+        [Theory]
+        [InlineData('a', 'a', 2)]
+        public void Complexity_should_behave_as_expected(char start, char end, double expected)
+        {
+            // Arrange
+            RangeBracketValue bracketValue = new(start, end);
+
+            // Act
+            double actual = bracketValue.Complexity;
+
+            // Assert
+            actual.Should()
+                  .Be(expected);
+        }
+
+        [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
+        public void Given_RangeBracketValue_Complexity_should_depends_on_start_and_end(RangeBracketValue bracketValue)
+        {
+            // Arrange
+            double expected = 1 + Math.Pow(2, bracketValue.End - bracketValue.Start);
+
+            // Act
+            double actual = bracketValue.Complexity;
+
+            // Assert
+            actual.Should()
+                  .Be(expected);
+        }
     }
 }
