@@ -10,6 +10,7 @@
     using FsCheck.Xunit;
 
     using System;
+    using System.Collections.Generic;
 
     using Xunit;
     using Xunit.Abstractions;
@@ -25,9 +26,7 @@
         [Fact]
         public void IsFilterExpression() => typeof(StringValueExpression).Should()
                                                                            .BeAssignableTo<FilterExpression>().And
-                                                                           .Implement<IEquatable<StringValueExpression>>().And
-                                                                           .HaveConstructor(new[] { typeof(string) }).And
-                                                                           .HaveProperty<string>("Value");
+                                                                           .Implement<IEquatable<StringValueExpression>>();
 
         [Fact]
         public void Ctor_Throws_ArgumentNullException_When_Argument_Is_Null()
@@ -63,20 +62,59 @@
         }
 
         [Property]
-        public Property Two_instances_are_equals_when_holding_same_values(NonEmptyString value)
+        public void Given_two_instances_that_hold_values_that_are_equals_Equals_shoud_return_true(NonEmptyString value)
         {
+            // Arrange
             StringValueExpression first = new(value.Item);
             StringValueExpression other = new(value.Item);
 
-            return first.Equals(other).ToProperty();
+            // Act
+            bool actual = first.Equals(other);
+
+            // Assert
+            actual.Should()
+                  .BeTrue();
+        }
+
+        public static IEnumerable<object[]> EqualsCases
+        {
+            get
+            {
+                yield return new object[]
+                {
+                    new StringValueExpression("True"),
+                    new OrExpression(new StringValueExpression("True"), new StringValueExpression("True")),
+                    true
+                };
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(EqualsCases))]
+        public void Equals_should_behave_as_expected(StringValueExpression stringValue, object obj, bool expected)
+        {
+            // Act
+            bool actual = stringValue.Equals(obj);
+
+            // Assert
+            actual.Should()
+                  .Be(expected);
         }
 
         [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
-        public Property Equals_is_commutative(StringValueExpression first, StringValueExpression second) => (first.Equals(second) == second.Equals(first)).ToProperty();
+        public Property Equals_should_be_commutative(StringValueExpression first, FilterExpression second)
+            => (first.Equals(second) == second.Equals(first)).ToProperty();
 
-        
+        [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
+        public Property Equals_should_be_reflexive(NonNull<StringValueExpression> expression)
+            => expression.Item.Equals(expression.Item).ToProperty();
+
+        [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
+        public Property Equals_should_be_symetric(NonNull<StringValueExpression> expression, NonNull<FilterExpression> otherExpression)
+            => (expression.Item.Equals(otherExpression.Item) == otherExpression.Item.Equals(expression.Item)).ToProperty();
+
         [Property]
-        public Property Given_two_ConstantExpressions_Equals_should_depends_on_string_input_only(NonWhiteSpaceString input)
+        public Property Given_two_StringValueExpresssions_Equals_should_depends_on_string_input_only(NonWhiteSpaceString input)
         {
             // Arrange
             StringValueExpression first = new(input.Get);
@@ -87,7 +125,7 @@
         }
 
         [Property(Arbitrary = new[] { typeof(ExpressionsGenerators) })]
-        public Property Given_ConstantExpression_GetComplexity_should_return_1(StringValueExpression constant) => (constant.Complexity == 1).ToProperty();
+        public Property Given_StringValueExpression_GetComplexity_should_return_1(StringValueExpression constant) => (constant.Complexity == 1).ToProperty();
 
         [Property]
         public void Given_StringValueExpression_and_TextExpression_are_based_on_same_value_IsEquivalentTo_should_be_true(NonWhiteSpaceString input)
@@ -102,6 +140,21 @@
             // Assert
             isEquivalent.Should()
                         .BeTrue($"{nameof(TextExpression)} was created from the same value");
+        }
+
+        [Property]
+        public void Given_StringValueExpression_when_GroupExpression_holds_an_expression_that_is_equals_to_that_StringValueExpression_Equals_should_return_true(NonWhiteSpaceString input)
+        {
+            // Arrange
+            StringValueExpression stringValue = new(input.Item);
+            GroupExpression group = new (new StringValueExpression(input.Item));
+
+            // Act
+            bool actual = stringValue.Equals(group);
+
+            // Assert
+            actual.Should()
+                  .BeTrue();
         }
     }
 }
